backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
psutil==7.2.1
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
sympy==1.14.0
torch==2.7.1+cu118
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
